+++ packages/flower_core/lib/flower_core.dart
@@ -6,3 +6,5 @@ export 'src/project_inspector.dart';
 export 'src/project_map.dart';
 export 'src/project_mapper.dart';
 export 'src/project_snapshot.dart';
+export 'src/project_symbol_index.dart';
+export 'src/project_symbol_indexer.dart';